Adapt OpenCL worker code to int64_t #19 * Appears to cause some performance loss in OpenCL * count_digits binary search tree dropped * Doesn't fix all overflow issues yet

+++ src/mine_DUCO_S1.c
@@ -36,15 +36,15 @@ long mine_DUCO_S1(
 }
 
 #ifndef NO_OPENCL
-long mine_DUCO_S1_OpenCL(
+int64_t mine_DUCO_S1_OpenCL(
     const unsigned char *input_prefix,
     int prefix_length,
     const unsigned char target_hexdigest[DUCO_S1_SIZE*2],
     int difficulty,
     worker_ctx *ctx)
 {
-    int maximum = 100*difficulty+1;
-    int correct_nonce;
+    int64_t maximum = 100*(int64_t)difficulty+1;
+    int64_t correct_nonce;
     init_OpenCL_worker_kernel(ctx, (const char*)input_prefix, (const char*)target_hexdigest);
 
     // We stagger GPU and CPU work here to avoid serial execution, and the GPU starts first
